+++ src/GRA.Data/Model/Questionnaire.cs
@@ -19,5 +19,10 @@ namespace GRA.Data.Model
         public ICollection<Question> Questions { get; set; }
 
         public bool IsLocked { get; set; }
+        public int? BadgeId { get; set; }
+        [MaxLength(255)]
+        public string BadgeName { get; set; }
+        [MaxLength(1000)]
+        public string BadgeNotificiationMessage { get; set; }
     }
 }