+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Add error message when trying to apply a role twice
 - Add trigger checking on login and account creation
 - Add names to program badges
+- Add badges to questionnaires
 
 ### Changed
 - Change join branch list to be populated with all branches